refactor: split types
