+++ src/requests/index.ts
@@ -1,2 +1,3 @@
 export * from "./requests";
 export * from "./helpers";
+export * from "./fetchData";
