+++ src/types/app-types.ts
@@ -1,0 +1,12 @@
+import { Country } from "./requests-types";
+
+export interface CustomButton {
+  key?: string | number;
+  label?: string;
+  onClick?: () => void;
+}
+
+export interface FindAllCountriesByLang {
+  countries: Country[] | null;
+  currentLangCode?: string;
+}
