+++ src/types/index.ts
@@ -1,1 +1,2 @@
-export * from "./types";
+export * from "./app-types";
+export * from "./requests-types";
