+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from "./convertObjIntoGraphQlRequest";
 export * from "./findAllCountriesByLang";
-export * from "./fetchData";
 export * from "./sortListByName";
